backend/requirements.txt
pymongo
pdfplumber
requests
python-dotenv
google-genai
